concord-common: move rest of copy methods to PathUtils
