+++ common/src/main/java/com/walmartlabs/concord/common/IOUtils.java
@@ -24,9 +24,8 @@ import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 
 import javax.validation.constraints.NotNull;
 import java.io.*;
-import java.nio.file.*;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.util.Collections;
+import java.nio.file.CopyOption;
+import java.nio.file.Path;
 import java.util.List;
 
 public final class IOUtils {
@@ -159,77 +158,44 @@ public final class IOUtils {
         }
     }
 
+    /**
+     * @deprecated use {@link PathUtils#copy(Path, Path)}
+     */
+    @Deprecated
     public static void copy(Path src, Path dst) throws IOException {
-        copy(src, dst, (String) null, null, new CopyOption[0]);
+        PathUtils.copy(src, dst);
     }
 
+    /**
+     * @deprecated use {@link PathUtils#copy(Path, Path, CopyOption...)}
+     */
+    @Deprecated
     public static void copy(Path src, Path dst, CopyOption... options) throws IOException {
-        copy(src, dst, (String) null, null, options);
+        PathUtils.copy(src, dst, options);
     }
 
+    /**
+     * @deprecated use {@link PathUtils#copy(Path, Path, String, CopyOption...)}
+     */
+    @Deprecated
     public static void copy(Path src, Path dst, String ignorePattern, CopyOption... options) throws IOException {
-        _copy(src, src, dst, toList(ignorePattern), null, options);
+        PathUtils.copy(src, dst, ignorePattern, options);
     }
 
+    /**
+     * @deprecated use {@link PathUtils#copy(Path, Path, String, FileVisitor, CopyOption...)}
+     */
+    @Deprecated
     public static void copy(Path src, Path dst, String skipContents, FileVisitor visitor, CopyOption... options) throws IOException {
-        _copy(src, src, dst, toList(skipContents), visitor, options);
+        PathUtils.copy(src, dst, skipContents, visitor, options);
     }
 
+    /**
+     * @deprecated use {@link PathUtils#copy(Path, Path, List, FileVisitor, CopyOption...)}
+     */
+    @Deprecated
     public static void copy(Path src, Path dst, List<String> skipContents, FileVisitor visitor, CopyOption... options) throws IOException {
-        _copy(src, src, dst, skipContents, visitor, options);
-    }
-
-    private static void _copy(Path root, Path src, Path dst, List<String> ignorePattern, FileVisitor visitor, CopyOption... options) throws IOException {
-        Files.walkFileTree(src, new SimpleFileVisitor<Path>() {
-            @Override
-            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) {
-                if (dir != src && anyMatch(src.relativize(dir).toString(), ignorePattern)) {
-                    return FileVisitResult.SKIP_SUBTREE;
-                }
-
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                if (file != src && anyMatch(src.relativize(file).toString(), ignorePattern)) {
-                    return FileVisitResult.CONTINUE;
-                }
-
-                Path a = file;
-                Path b = dst.resolve(src.relativize(file));
-
-                Path parent = b.getParent();
-                if (!Files.exists(parent)) {
-                    Files.createDirectories(parent);
-                }
-
-                if (Files.isSymbolicLink(file)) {
-                    Path link = Files.readSymbolicLink(file);
-                    Path target = file.getParent().resolve(link).normalize();
-
-                    if (!target.startsWith(root)) {
-                        throw new IOException("Symlinks outside the base directory are not supported: " + file + " -> " + target);
-                    }
-
-                    if (Files.notExists(target)) {
-                        // missing target
-                        return FileVisitResult.CONTINUE;
-                    }
-
-                    Files.createSymbolicLink(b, link);
-                    return FileVisitResult.CONTINUE;
-                }
-
-                Files.copy(a, b, options);
-
-                if (visitor != null) {
-                    visitor.visit(a, b);
-                }
-
-                return FileVisitResult.CONTINUE;
-            }
-        });
+        PathUtils.copy(src, dst, skipContents, visitor, options);
     }
 
     /**
@@ -245,7 +211,7 @@ public final class IOUtils {
      */
     @Deprecated
     public static List<String> grep(String pattern, InputStream in) throws IOException {
-       return GrepUtils.grep(pattern, in);
+        return GrepUtils.grep(pattern, in);
     }
 
     /**
@@ -280,22 +246,6 @@ public final class IOUtils {
     @Deprecated
     public static Path assertInPath(@NotNull Path parent, @NotNull String child) throws IOException {
         return PathUtils.assertInPath(parent, child);
-    }
-
-    private static List<String> toList(String entry) {
-        if (entry == null) {
-            return Collections.emptyList();
-        }
-
-        return Collections.singletonList(entry);
-    }
-
-    private static boolean anyMatch(String what, List<String> patterns) {
-        if (patterns == null) {
-            return false;
-        }
-
-        return patterns.stream().anyMatch(what::matches);
     }
 
     private IOUtils() {
