+++ common/src/main/java/com/walmartlabs/concord/common/PathUtils.java
@@ -28,6 +28,8 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.*;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.util.Collections;
+import java.util.List;
 
 public final class PathUtils {
 
@@ -133,5 +135,94 @@ public final class PathUtils {
     }
 
     private PathUtils() {
+    }
+
+    public static void copy(Path src, Path dst) throws IOException {
+        copy(src, dst, (String) null, null, new CopyOption[0]);
+    }
+
+    public static void copy(Path src, Path dst, CopyOption... options) throws IOException {
+        copy(src, dst, (String) null, null, options);
+    }
+
+    public static void copy(Path src, Path dst, String ignorePattern, CopyOption... options) throws IOException {
+        _copy(src, src, dst, toList(ignorePattern), null, options);
+    }
+
+    public static void copy(Path src, Path dst, String skipContents, FileVisitor visitor, CopyOption... options) throws IOException {
+        _copy(src, src, dst, toList(skipContents), visitor, options);
+    }
+
+    public static void copy(Path src, Path dst, List<String> skipContents, FileVisitor visitor, CopyOption... options) throws IOException {
+        _copy(src, src, dst, skipContents, visitor, options);
+    }
+
+    private static void _copy(Path root, Path src, Path dst, List<String> ignorePattern, FileVisitor visitor, CopyOption... options) throws IOException {
+        Files.walkFileTree(src, new SimpleFileVisitor<Path>() {
+            @Override
+            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) {
+                if (dir != src && anyMatch(src.relativize(dir).toString(), ignorePattern)) {
+                    return FileVisitResult.SKIP_SUBTREE;
+                }
+
+                return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                if (file != src && anyMatch(src.relativize(file).toString(), ignorePattern)) {
+                    return FileVisitResult.CONTINUE;
+                }
+
+                Path a = file;
+                Path b = dst.resolve(src.relativize(file));
+
+                Path parent = b.getParent();
+                if (!Files.exists(parent)) {
+                    Files.createDirectories(parent);
+                }
+
+                if (Files.isSymbolicLink(file)) {
+                    Path link = Files.readSymbolicLink(file);
+                    Path target = file.getParent().resolve(link).normalize();
+
+                    if (!target.startsWith(root)) {
+                        throw new IOException("Symlinks outside the base directory are not supported: " + file + " -> " + target);
+                    }
+
+                    if (Files.notExists(target)) {
+                        // missing target
+                        return FileVisitResult.CONTINUE;
+                    }
+
+                    Files.createSymbolicLink(b, link);
+                    return FileVisitResult.CONTINUE;
+                }
+
+                Files.copy(a, b, options);
+
+                if (visitor != null) {
+                    visitor.visit(a, b);
+                }
+
+                return FileVisitResult.CONTINUE;
+            }
+        });
+    }
+
+    private static List<String> toList(String entry) {
+        if (entry == null) {
+            return Collections.emptyList();
+        }
+
+        return Collections.singletonList(entry);
+    }
+
+    private static boolean anyMatch(String what, List<String> patterns) {
+        if (patterns == null) {
+            return false;
+        }
+
+        return patterns.stream().anyMatch(what::matches);
     }
 }
