+++ common/src/test/java/com/walmartlabs/concord/common/PathUtilsTest.java
@@ -23,12 +23,12 @@ package com.walmartlabs.concord.common;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PathUtilsTest {
 
@@ -47,5 +47,70 @@ public class PathUtilsTest {
 
         Path p = assertDoesNotThrow(() -> PathUtils.assertInPath(parent, "odd/../but/valid"));
         assertEquals("/parent/path/but/valid", p.toString());
+    }
+
+
+    @Test
+    public void testCopy() throws Exception {
+        Path src = Files.createTempDirectory("test");
+        Path dst = Files.createTempDirectory("test");
+
+        // ---
+
+        Path nestedDir = src.resolve("a/b");
+        Files.createDirectories(nestedDir);
+
+        Path srcFile = nestedDir.resolve("c.txt");
+        Files.createFile(srcFile);
+
+        // ---
+
+        PathUtils.copy(src, dst);
+        assertTrue(Files.exists(dst.resolve("a/b/c.txt")));
+    }
+
+    @Test
+    public void testSymlinks() throws Exception {
+        Path src = Files.createTempDirectory("test");
+
+        Path aFile = src.resolve("a");
+        Files.write(aFile, "hello".getBytes(), StandardOpenOption.CREATE);
+
+        Path xDir = src.resolve("x");
+        Files.createDirectories(xDir);
+
+        Path bLink = xDir.resolve("b");
+        Files.createSymbolicLink(bLink, aFile);
+
+        // ---
+
+        Path dst = Files.createTempDirectory("test");
+
+        // ---
+
+        PathUtils.copy(src, dst);
+
+        // ---
+
+        assertTrue(Files.isSymbolicLink(dst.resolve("x").resolve("b")));
+        assertTrue(Files.isRegularFile(dst.resolve("x").resolve("b")));
+    }
+
+    @Test
+    public void testExternalSymlinks() throws Exception {
+        Path src = Files.createTempDirectory("test");
+
+        Path link = src.resolve("a");
+        Path target = Paths.get("../../../etc/passwd");
+        Files.createSymbolicLink(link, target);
+
+        // ---
+
+        Path dst = Files.createTempDirectory("test");
+
+        // ---
+
+        Exception e = assertThrows(IOException.class, () -> PathUtils.copy(src, dst));
+        assertTrue(e.getMessage().contains("Symlinks outside the base directory are not supported"));
     }
 }
