+++ it/server/src/test/java/com/walmartlabs/concord/it/server/GitHubNonOrgEventIt.java
@@ -22,7 +22,7 @@ package com.walmartlabs.concord.it.server;
 
 import com.walmartlabs.concord.client2.*;
 import com.walmartlabs.concord.client2.ProcessListFilter;
-import com.walmartlabs.concord.common.IOUtils;
+import com.walmartlabs.concord.common.PathUtils;
 import com.walmartlabs.concord.it.common.GitHubUtils;
 import org.eclipse.jgit.api.Git;
 import org.junit.jupiter.api.Test;
@@ -45,7 +45,7 @@ public class GitHubNonOrgEventIt extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(GitHubNonOrgEventIt.class.getResource("githubNonRepoEvent").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         try (Git repo = Git.init().setInitialBranch("master").setDirectory(tmpDir.toFile()).call()) {
             repo.add().addFilepattern(".").call();
