+++ it/server/src/test/java/com/walmartlabs/concord/it/server/MavenRepoIT.java
@@ -21,7 +21,7 @@ package com.walmartlabs.concord.it.server;
  */
 
 import com.walmartlabs.concord.client2.*;
-import com.walmartlabs.concord.common.IOUtils;
+import com.walmartlabs.concord.common.PathUtils;
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
@@ -49,7 +49,7 @@ public class MavenRepoIT extends AbstractServerIT {
         Path mvnDirectory = localMavenRepo.resolve("com/walmartlabs/concord/mvn-concord/0.0.1");
 
         Path src = Paths.get(MavenRepoIT.class.getResource("mvnRepoFiles").toURI());
-        IOUtils.copy(src, mvnDirectory, StandardCopyOption.REPLACE_EXISTING);
+        PathUtils.copy(src, mvnDirectory, StandardCopyOption.REPLACE_EXISTING);
 
         String url = "mvn://com.walmartlabs.concord:mvn-concord:zip";
         String projectName = "project_" + randomString();
