+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ProcessContainerIT.java
@@ -22,7 +22,6 @@ package com.walmartlabs.concord.it.server;
 
 import com.walmartlabs.concord.client2.ProcessEntry;
 import com.walmartlabs.concord.client2.StartProcessResponse;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -44,7 +43,7 @@ public class ProcessContainerIT extends AbstractServerIT {
     public void test() throws Exception {
         Path src = Paths.get(ProcessContainerIT.class.getResource("processContainer").toURI());
         Path dst = PathUtils.createTempDir("test");
-        IOUtils.copy(src, dst);
+        PathUtils.copy(src, dst);
 
         Path concordYml = dst.resolve("concord.yml");
         String s = new String(Files.readAllBytes(concordYml));
