+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ProjectIT.java
@@ -21,7 +21,6 @@ package com.walmartlabs.concord.it.server;
  */
 
 import com.walmartlabs.concord.client2.*;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
 import com.walmartlabs.concord.sdk.Constants;
 import org.eclipse.jgit.api.Git;
@@ -44,7 +43,7 @@ public class ProjectIT extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(ProjectIT.class.getResource("project").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         try (Git repo = initRepo(tmpDir)) {
             repo.add().addFilepattern(".").call();
@@ -82,7 +81,7 @@ public class ProjectIT extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(ProjectIT.class.getResource("projectEntryPoint").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         try (Git repo = initRepo(tmpDir)) {
             repo.add().addFilepattern(".").call();
@@ -112,7 +111,7 @@ public class ProjectIT extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(ProjectIT.class.getResource("project").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         try (Git repo = initRepo(tmpDir)) {
             repo.add().addFilepattern(".").call();
@@ -121,7 +120,7 @@ public class ProjectIT extends AbstractServerIT {
             // commit-1
             PathUtils.deleteRecursively(tmpDir.resolve("processes"));
             src = new File(ProjectIT.class.getResource("project-commit-id").toURI());
-            IOUtils.copy(src.toPath().resolve("1"), tmpDir);
+            PathUtils.copy(src.toPath().resolve("1"), tmpDir);
             Ref result = repo.checkout().setName("test-branch").setCreateBranch(true).call();
             assertNotNull(result);
 
@@ -133,7 +132,7 @@ public class ProjectIT extends AbstractServerIT {
             result = repo.checkout().setName("master").call();
             PathUtils.deleteRecursively(tmpDir.resolve("processes"));
             src = new File(ProjectIT.class.getResource("project-commit-id").toURI());
-            IOUtils.copy(src.toPath().resolve("2"), tmpDir);
+            PathUtils.copy(src.toPath().resolve("2"), tmpDir);
             assertNotNull(result);
             repo.add().addFilepattern(".").call();
             repo.commit().setMessage("commit-2").call();
@@ -165,7 +164,7 @@ public class ProjectIT extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(ProjectIT.class.getResource("project").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         String tag = "tag_for_testing";
 
@@ -176,7 +175,7 @@ public class ProjectIT extends AbstractServerIT {
             // commit-1
             PathUtils.deleteRecursively(tmpDir.resolve("processes"));
             src = new File(ProjectIT.class.getResource("project-commit-id").toURI());
-            IOUtils.copy(src.toPath().resolve("1"), tmpDir);
+            PathUtils.copy(src.toPath().resolve("1"), tmpDir);
             repo.add().addFilepattern(".").call();
             repo.commit().setMessage("commit-1").call();
 
@@ -185,7 +184,7 @@ public class ProjectIT extends AbstractServerIT {
             // commit-2
             PathUtils.deleteRecursively(tmpDir.resolve("processes"));
             src = new File(ProjectIT.class.getResource("project-commit-id").toURI());
-            IOUtils.copy(src.toPath().resolve("2"), tmpDir);
+            PathUtils.copy(src.toPath().resolve("2"), tmpDir);
             repo.add().addFilepattern(".").call();
             repo.commit().setMessage("commit-2").call();
         }
@@ -215,7 +214,7 @@ public class ProjectIT extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(ProjectIT.class.getResource("project-triggers").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         try (Git repo = initRepo(tmpDir)) {
             repo.add().addFilepattern(".").call();
@@ -252,7 +251,7 @@ public class ProjectIT extends AbstractServerIT {
         Path tmpDir = createTempDir();
 
         File src = new File(ProjectIT.class.getResource("repositoryValidation").toURI());
-        IOUtils.copy(src.toPath(), tmpDir);
+        PathUtils.copy(src.toPath(), tmpDir);
 
         try (Git repo = initRepo(tmpDir)) {
             repo.add().addFilepattern(".").call();
@@ -286,7 +285,7 @@ public class ProjectIT extends AbstractServerIT {
             Path tmpDir = createTempDir();
 
             File src = new File(ProjectIT.class.getResource("repositoryValidationEmptyFlow").toURI());
-            IOUtils.copy(src.toPath(), tmpDir);
+            PathUtils.copy(src.toPath(), tmpDir);
 
             try (Git repo = initRepo(tmpDir)) {
                 repo.add().addFilepattern(".").call();
@@ -321,7 +320,7 @@ public class ProjectIT extends AbstractServerIT {
             Path tmpDir = createTempDir();
 
             File src = new File(ProjectIT.class.getResource("repositoryValidationEmptyForm").toURI());
-            IOUtils.copy(src.toPath(), tmpDir);
+            PathUtils.copy(src.toPath(), tmpDir);
 
             try (Git repo = initRepo(tmpDir)) {
                 repo.add().addFilepattern(".").call();
@@ -356,7 +355,7 @@ public class ProjectIT extends AbstractServerIT {
             Path tmpDir = createTempDir();
 
             File src = new File(ProjectIT.class.getResource("ProcessDisabledRepo").toURI());
-            IOUtils.copy(src.toPath(), tmpDir);
+            PathUtils.copy(src.toPath(), tmpDir);
 
             try (Git repo = initRepo(tmpDir)) {
                 repo.add().addFilepattern(".").call();
