+++ repository/src/main/java/com/walmartlabs/concord/repository/GitCliRepositoryProvider.java
@@ -20,7 +20,6 @@ package com.walmartlabs.concord.repository;
  * =====
  */
 
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -80,7 +79,7 @@ public class GitCliRepositoryProvider implements RepositoryProvider {
         List<String> allIgnorePatterns = new ArrayList<>();
         allIgnorePatterns.add(GIT_FILES);
         allIgnorePatterns.addAll(ignorePatterns);
-        IOUtils.copy(src, dst, allIgnorePatterns, snapshot, StandardCopyOption.REPLACE_EXISTING);
+        PathUtils.copy(src, dst, allIgnorePatterns, snapshot, StandardCopyOption.REPLACE_EXISTING);
         return snapshot;
     }
 }
